% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_func.R
\name{center}
\alias{center}
\title{Matrix centering}
\usage{
center(x, subset = NULL, byColumn = TRUE, na.rm = FALSE, nhead = 51)
}
\arguments{
\item{x}{A numeric matrix or data frame.}

\item{subset}{Encoded value of an index vector indicating the subset of individuals to consider. 
Default, NULL, all individuals are considered.}

\item{byColumn}{A logical value indicating whether the input data is centered by column or row.
Default, TRUE, centering by column. Constant variables across samples are removed. 
If FALSE, centering and scaling by row. Constant samples across variables are removed.}

\item{na.rm}{A logical value indicating NA values should be removed. Default, FALSE, NA set to 0.}

\item{nhead}{Deprecated}
}
\value{
A centered matrix with 0-mean per column (by default).
}
\description{
Center matrix columns or rows to 0
}
