% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiblock_aggregate.R
\name{colmeans}
\alias{colmeans}
\title{Column means, deprecated

Column means of a matrix}
\usage{
colmeans(x)
}
\arguments{
\item{x}{A numeric matrix}
}
\value{
Column means of x
}
\description{
Column means, deprecated

Column means of a matrix
}
