% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_func.R
\name{colNames}
\alias{colNames}
\title{Colnames}
\usage{
colNames(x)
}
\arguments{
\item{x}{A matrix, a data frame, or a list of matrices or data frames}
}
\value{
Column names of x
}
\description{
Get column names of a matrix
}
