% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serv_scudo.R
\name{computeWeights}
\alias{computeWeights}
\title{computeWeights

Weight most and least regulated features in original dataMatrix accordingly to ranking}
\usage{
computeWeights(
  expressionData,
  indexMatrix,
  top = 2,
  bottom = 2,
  topWeight = 10,
  bottomWeight = 0.1
)
}
\arguments{
\item{expressionData}{A matrix or data frame, samples in rows and features in columns}

\item{indexMatrix}{rank matrix of same dimension of expressionData matrix}

\item{top}{which most regulated genes we consider}

\item{bottom}{which least regulated genes we consider}

\item{topWeight}{weight associated to top elements}

\item{bottomWeight}{weight associated to bottom elements}
}
\value{
outWeiight matrix of the same dimension of expressionData
}
\description{
computeWeights

Weight most and least regulated features in original dataMatrix accordingly to ranking
}
