% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_func.R
\name{crossAggregate}
\alias{crossAggregate}
\title{Cross aggregate}
\usage{
crossAggregate(conns, expr, wait = F, async = T)
}
\arguments{
\item{conns}{A list of DSConnection-class.}

\item{expr}{An encoded expression to evaluate.}

\item{wait}{See DSI::datashield.aggregate options. Default: FALSE.}

\item{async}{See DSI::datashield.aggregate options. Default: TRUE.}
}
\description{
Call datashield.aggregate on remote servers.
}
