% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_func.R
\name{crossAggregate2}
\alias{crossAggregate2}
\title{Cross aggregate through two-layer connection}
\usage{
crossAggregate2(conns, expr, async = T)
}
\arguments{
\item{conns}{A list of DSConnection-class.}

\item{expr}{An encoded expression to evaluate.}

\item{async}{See DSI::datashield.aggregate options. Default: TRUE.}
}
\description{
Call datashield.aggregate on remote servers through two-layer connection.
}
