% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_func.R
\name{crossAggregateDual}
\alias{crossAggregateDual}
\title{Cross aggregate through two-layer connection}
\usage{
crossAggregateDual(conns, expr, async = T)
}
\arguments{
\item{conns}{A list of Opal connections.}

\item{expr}{An encoded expression to evaluate. This is restricted to pushToDscFD.}

\item{async}{See DSI::datashield.aggregate options. Default, TRUE.}
}
\description{
Call datashield.aggregate on remote servers through two-layer connection.
}
