% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_func.R
\name{crossAggregatePrimal}
\alias{crossAggregatePrimal}
\title{Cross aggregate}
\usage{
crossAggregatePrimal(conns, expr, async = T)
}
\arguments{
\item{conns}{A list of DSConnection-class.}

\item{expr}{An encoded expression to evaluate.}

\item{async}{See DSI::datashield.aggregate options. Default: TRUE.}
}
\description{
Call datashield.aggregate on remote servers.
}
