% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_func.R
\name{crossAssign}
\alias{crossAssign}
\title{Cross assign}
\usage{
crossAssign(conns, symbol, value, value.call, async = T)
}
\arguments{
\item{conns}{A list of Opal connections.}

\item{symbol}{Name of an R symbol.}

\item{value}{A variable name or an R expression with allowed assign function calls.}

\item{value.call}{A logical value, TRUE if value is function call, FALSE if value is a variable name.}

\item{async}{See DSI::datashield.assign options. Default: TRUE.}
}
\description{
Call datashield.assign on remote servers.
}
