% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_func.R
\name{crossLogin}
\alias{crossLogin}
\title{Cross login

Cross login # rather on client side}
\usage{
crossLogin(logins)
}
\arguments{
\item{logins}{An encoded dataframe with server, url, user, password, and driver fields.}
}
\description{
Cross login

Cross login # rather on client side
}
