% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_func.R
\name{crossProd}
\alias{crossProd}
\title{Matrix cross product}
\usage{
crossProd(x, y = NULL)
}
\arguments{
\item{x}{A numeric matrix}
}
\value{
t(x) \%*\% x
}
\description{
Calculates the cross product t(x) \%*\% x
}
