% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_func.R
\name{crossProdrm}
\alias{crossProdrm}
\title{Matrix cross product}
\usage{
crossProdrm(x, y = NULL)
}
\arguments{
\item{x}{A numeric matrix}

\item{y}{An encoded matrix. Default, NULL, y = x.}
}
\value{
t(x) \%*\% y
}
\description{
Calculates the cross product t(x) \%*\% y
}
