% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{.cleanup}
\alias{.cleanup}
\title{Remove objects from the current workspace}
\usage{
.cleanup(initial, start = ".GlobalEnv")
}
\arguments{
\item{start}{a character the environment name where to start (default .GlobalEnv)}

\item{what}{a list of  environments and their respective objects - the output of a previous call to ls.all}
}
\description{
This function removes objects from the current workspace
See \code{dsSwissKnife:::.cleanup}.
}
\keyword{internal}
