% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{.encode.arg}
\alias{.encode.arg}
\title{Encode function argument}
\usage{
.encode.arg(some.object, serialize.it = TRUE, digits = 20)
}
\arguments{
\item{some.object}{the object to be encoded}
}
\value{
encoded text with offending characters replaced by strings
}
\description{
Serialize to JSON, then encode base64,
then replace '+', '/' and '=' in the result in order to play nicely with the
opal entry. Used to encode non-scalar function arguments prior to sending to
the opal server. There is a corresponding function in the server package
called \code{.decode.arg}. See \code{dsSwissKnifeClient:::.encode.arg}.
}
\keyword{internal}
