% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{.lock.unlock}
\alias{.lock.unlock}
\title{Locks or unlocks bindings in environments}
\usage{
.lock.unlock(what, func)
}
\arguments{
\item{what}{a list of  environments and their respective objects - the output of ls.all above.}

\item{func}{a function, either lockBinding or unlockBinding.}
}
\description{
Locks or unlocks bindings in environments.
See \code{dsSwissKnife:::.lock.unlock}.
}
\keyword{internal}
