% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{.login}
\alias{.login}
\title{Wrapper of datashield.login}
\usage{
.login(logins)
}
\arguments{
\item{logins}{A dataframe of login information. See \code{datashield.login}.}
}
\value{
Object(s) of class OpalConnection
}
\description{
This function ensures \code{datashield.login} to all the
servers without error of simultaneously connecting to the same server.
}
\keyword{internal}
