% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_func.R
\name{.partitionMatrix}
\alias{.partitionMatrix}
\title{Matrix partition}
\usage{
.partitionMatrix(x, seprow, sepcol = seprow)
}
\arguments{
\item{x}{A matrix}

\item{seprow}{A numeric vectors indicating sizes of blocks in rows}

\item{sepcol}{A numeric vectors indicating sizes of blocks in columns}
}
\value{
List of blocks
}
\description{
Partition a matrix into blocks
}
\keyword{internal}
