% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_func.R
\name{.rebuildMatrix}
\alias{.rebuildMatrix}
\title{Matrix reconstruction}
\usage{
.rebuildMatrix(matblocks, mc.cores = 1)
}
\arguments{
\item{matblocks}{List of lists of matrix blocks, obtained from
.partitionMatrix.}

\item{mc.cores}{Number of cores for parallel computing. Default, 1.}
}
\value{
The reconstructed matrix.
}
\description{
Rebuild a matrix from its partition.
}
\keyword{internal}
