% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_func.R
\name{.rebuildMatrixEnc}
\alias{.rebuildMatrixEnc}
\title{Symmetric matrix reconstruction}
\usage{
.rebuildMatrixEnc(blocks, mc.cores = 1)
}
\arguments{
\item{blocks}{List of list of encoded matrix blocks}

\item{mc.cores}{Number of cores for parallel computing. Default: 1}
}
\value{
The complete symmetric matrix
}
\description{
Rebuild a matrix from its encoded partition
}
\keyword{internal}
