% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_func.R
\name{.sumMatrices}
\alias{.sumMatrices}
\title{Sum matrices}
\usage{
.sumMatrices(dsc = NULL)
}
\arguments{
\item{dsc}{A list of big memory descriptions}
}
\value{
Sum of x and those stored in dsc
}
\description{
Compute the sum of a matrix and those stored in bigmemory
}
\keyword{internal}
