% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_func.R
\name{dsDim}
\alias{dsDim}
\title{Matrix dimension}
\usage{
dsDim(x)
}
\arguments{
\item{x}{A matrix, a data frame, or a list of matrices or data frames}
}
\value{
Dimension of x
}
\description{
Dimension of a data frame or matrix
}
