% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_func.R
\name{dsFactor}
\alias{dsFactor}
\title{Factor coercion}
\usage{
dsFactor(x)
}
\arguments{
\item{x}{A variable of a data frame in form of 'data.frame$variable'.}
}
\value{
An object of class \code{factor}
}
\description{
Factor coercion of a categorical variable in a data frame
}
