% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_func.R
\name{dsLevels}
\alias{dsLevels}
\title{Levels attributes}
\usage{
dsLevels(x)
}
\arguments{
\item{x}{A variable of a data frame in form of 'data.frame$variable'}
}
\value{
Levels of x
}
\description{
Return levels of a categorical variable in a data frame
}
