% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_func.R
\name{dsRange}
\alias{dsRange}
\title{Range of a variable in a data frame}
\usage{
dsRange(x)
}
\arguments{
\item{x}{A variable of a data frame in form of 'data.frame$variable'.}
}
\value{
Approximate range of x
}
\description{
Return random approximated range of a numeric variable in a
data frame
}
