% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_func.R
\name{dsRank}
\alias{dsRank}
\title{Ranking, deprecated}
\usage{
dsRank(x)
}
\arguments{
\item{x}{A matrix or data frame, samples in rows and features in columns}
}
\value{
Ranking of features
}
\description{
Ranking of features in each sample
}
