% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_func.R
\name{loadings}
\alias{loadings}
\title{Loadings}
\usage{
loadings(x, y, operator = "crossprod")
}
\arguments{
\item{x}{A numeric matrix}

\item{y}{A numeric matrix for the new basis of the same nrow to x.}

\item{operator}{An operation to compute the loadings (\code{crossprod}, \code{cor}). Default, \code{crossprod}}
}
\value{
operator(x, y)
}
\description{
Loadings of features in a new feature basis
}
