% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_func.R
\name{matrix2DscServer}
\alias{matrix2DscServer}
\title{Bigmemory description of a matrix}
\usage{
matrix2DscServer(value)
}
\arguments{
\item{value}{Encoded value of a matrix}
}
\value{
Bigmemory description of the given matrix
}
\description{
Bigmemory description of a matrix
}
