% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_func.R
\name{push}
\alias{push}
\title{Push

Push the output of a function call to other servers}
\usage{
push(opal, symbol, value, value.call, variables = NULL, wait = F, async = T)
}
\arguments{
\item{opal}{A list of opal objects.}
}
\description{
Push

Push the output of a function call to other servers
}
