% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_func.R
\name{pushSymmMatrixServer}
\alias{pushSymmMatrixServer}
\title{Push a symmetric matrix}
\usage{
pushSymmMatrixServer(value)
}
\arguments{
\item{value}{An encoded value to be pushed}
}
\value{
Description of the pushed value
}
\description{
Push symmetric matrix data into the federated server
}
