% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_func.R
\name{pushToDsc}
\alias{pushToDsc}
\title{Bigmemory description of a pushed object}
\usage{
pushToDsc(conns, symbol, async = T)
}
\arguments{
\item{conns}{A one-element list of DSConnection-class.}

\item{symbol}{Name of an object to be pushed.}

\item{async}{See DSI::datashield.aggregate options. Default: TRUE.}
}
\value{
Bigmemory description of the pushed object on conns
}
\description{
Bigmemory description of a pushed object
}
