% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_func.R
\name{pushToDscFD}
\alias{pushToDscFD}
\title{Bigmemory description of a pushed object}
\usage{
pushToDscFD(conns, object, async = T)
}
\arguments{
\item{conns}{A list of Opal connections. Only one connection is accepted.}

\item{object}{The object to be pushed.}

\item{async}{See DSI::datashield.aggregate options. Default, TRUE.}
}
\value{
Bigmemory description of the pushed object on \code{conns}.
}
\description{
Bigmemory description of a pushed object
}
