% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_func.R
\name{pushToDscMate}
\alias{pushToDscMate}
\title{Bigmemory description of a pushed object}
\usage{
pushToDscMate(conns, symbol, sourcename, async = T)
}
\arguments{
\item{conns}{A list of DSConnection-classes.}

\item{symbol}{Name of the object to be pushed.}

\item{sourcename}{Name of the pushed object source.}

\item{async}{See DSI::datashield.aggregate options. Default: TRUE.}
}
\value{
Bigmemory description of the pushed object on conns
}
\description{
Bigmemory description of a pushed object
}
