% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_func.R
\name{pushValue}
\alias{pushValue}
\title{Memory description}
\usage{
pushValue(value, name)
}
\arguments{
\item{value}{A symmetric matrix}
}
\value{
Bigmemory description of the given matrix
}
\description{
Push a symmetric matrix into some R session by its bigmemory description in the session.
}
