% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiblock_assign.R
\name{randomData}
\alias{randomData}
\title{Random matrix

Create a random data matrix}
\usage{
randomData(nrow, ncol)
}
\arguments{
\item{nrow}{Number of rows}

\item{ncol}{Number of colums}
}
\value{
A matrix of dimension (nrow, ncol)
}
\description{
Random matrix

Create a random data matrix
}
