% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_func.R
\name{rebuildMatrix}
\alias{rebuildMatrix}
\title{Matrix rebuild}
\usage{
rebuildMatrix(blocks)
}
\arguments{
\item{blocks}{List of list of encoded matrix blocks, obtained from crossProd or tcrossProd}
}
\value{
The complete matrix
}
\description{
Rebuild a matrix from its partition
}
\keyword{internal}
