% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_func.R
\name{rebuildMatrixVar}
\alias{rebuildMatrixVar}
\title{Symmetric matrix reconstruction}
\usage{
rebuildMatrixVar(symbol, len, mc.cores = 1)
}
\arguments{
\item{symbol}{Generic variable name}

\item{len}{Length of list of lists of matrix blocks. Variables were generated as symbol11, symbol12, etc.}

\item{mc.cores}{Number of cores for parallel computing. Default: 1}
}
\value{
The complete symmetric matrix
}
\description{
Rebuild a symmetric matrix from its partition through variables in the environment
}
