% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiblock_aggregate.R
\name{rowmeans}
\alias{rowmeans}
\title{Row means

Row means of linear transformations of a matrix}
\usage{
rowmeans(x, y = NULL)
}
\arguments{
\item{x}{A numeric matrix}

\item{y}{A list of symmetric matrices of dimension (ncol(x), ncol(x))}
}
\value{
Row means of x if y = NULL, or row means x %*% yy for each matrix yy in y otherwise
}
\description{
Row means

Row means of linear transformations of a matrix
}
