% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_func.R
\name{setRowNames}
\alias{setRowNames}
\title{Row names}
\usage{
setRowNames(x, row.names)
}
\arguments{
\item{x}{A matrix or a data frame}

\item{row.names}{An encoded vector of names with the length of nrow(x)}
}
\value{
Matrix with row names
}
\description{
Assign row names to a matrix
}
