% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_func.R
\name{setRowNames.rm}
\alias{setRowNames.rm}
\title{Assign rownames}
\usage{
setRowNames.rm(x, row.names, envir = .GlobalEnv)
}
\arguments{
\item{x}{A numeric matrix}

\item{row.names}{An encoded vector of names with the length of nrow(x), or a
name of variable in \code{envir}.}

\item{envir}{A data frame or environment where row.names will be queried if
it is a variable name. Default, \code{.GlobalEnv}}
}
\value{
Matrix with rownames
}
\description{
Assign row names to a matrix
}
\keyword{internal}
