% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_func.R
\name{tcrossProd}
\alias{tcrossProd}
\title{Matrix cross product}
\usage{
tcrossProd(x, y = NULL, chunk = 500)
}
\arguments{
\item{x}{A numeric matrix}

\item{y}{A list of numeric matrices. Default, y = x.}
}
\value{
\code{x \%*\% t(y)}
}
\description{
Calculates the cross product x \%*\% t(y)
}
