% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_func.R
\name{tripleProd}
\alias{tripleProd}
\title{Matrix triple product}
\usage{
tripleProd(x, pids)
}
\arguments{
\item{x}{A numeric matrix}

\item{y}{A list of symmetric numeric matrices of dimension (ncol(x), ncol(x))}
}
\value{
List of x \%*\% y \%*\% t(x)
}
\description{
Calculate the triple product x \%*\% y \%*\% t(x)
}
