% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_func.R
\name{tripleProdChunk}
\alias{tripleProdChunk}
\title{Matrix triple product}
\usage{
tripleProdChunk(x, mate, chunk = 500L, mc.cores = 1)
}
\arguments{
\item{x}{A list of numeric matrices.}

\item{mate}{Mate server name, from which y was pushed.}

\item{chunk}{Size of chunks into what the resulting matrix is partitioned.
Default, 500L.}

\item{mc.cores}{Number of cores for parallel computing. Default, 1.}
}
\value{
\code{x \%*\% t(y)}
}
\description{
Calculate the triple product x \%*\% y \%*\% t(x).
}
