% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiblock_assign.R
\name{trscale}
\alias{trscale}
\title{Matrix scaling

Scale a matrix to trace of 1}
\usage{
trscale(x, na.rm = TRUE)
}
\arguments{
\item{x}{A numeric matrix or data frame.}
}
\value{
A scaled matrix with tr(xx') = 1
}
\description{
Matrix scaling

Scale a matrix to trace of 1
}
