% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_func.R
\name{center}
\alias{center}
\title{Matrix centering

Center matrix columns to 0}
\usage{
center(x, na.rm = FALSE)
}
\arguments{
\item{x}{A numeric matrix or data frame.}

\item{na.rm}{A logical value indicating NA values should be removed. Default, FALSE, NA set to 0.}
}
\value{
A centered matrix with column mean = 0
}
\description{
Matrix centering

Center matrix columns to 0
}
