% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiblock_assign.R
\name{center}
\alias{center}
\title{Matrix centering

Center matrix columns to 0}
\usage{
center(x, na.rm = TRUE)
}
\arguments{
\item{x}{A numeric matrix or data frame.}
}
\value{
A centered matrix with column mean = 0
}
\description{
Matrix centering

Center matrix columns to 0
}
