% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_func.R
\name{center}
\alias{center}
\title{Matrix centering}
\usage{
center(x, subset = NULL, byColumn = TRUE, na.rm = FALSE, nhead = 51)
}
\arguments{
\item{x}{A numeric matrix or data frame.}

\item{subset}{Encoded value of an index vector indicating the subset of individuals to consider. 
Default, NULL, all individuals are considered.}

\item{na.rm}{A logical value indicating NA values should be removed. Default, FALSE, NA set to 0.}
}
\value{
A centered matrix with column mean = 0
}
\description{
Center matrix columns to 0
}
