% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiblock_aggregate.R
\name{crossAggregate}
\alias{crossAggregate}
\title{Cross aggregate

Cross aggregate # rather on client side}
\usage{
crossAggregate(opal, expr, wait = F, async = T)
}
\arguments{
\item{opal}{An opal object or list of opal objects.}

\item{expr}{An encoded expression to evaluate.}

\item{wait}{See DSI::datashield.aggreate options. Default: FALSE.}

\item{async}{See DSI::datashield.aggreate options. Default: TRUE.}
}
\description{
Cross aggregate

Cross aggregate # rather on client side
}
