% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiblock_aggregate.R
\name{crossAssign}
\alias{crossAssign}
\title{Cross assign

Cross assign # rather on client side}
\usage{
crossAssign(
  opal,
  symbol,
  value,
  value.call,
  variables = NULL,
  wait = F,
  async = T
)
}
\arguments{
\item{opal}{An opal object or list of opal objects.}

\item{symbol}{Name of an R symbol.}

\item{value}{A variable name or an R epxression with allowed assign function calls.}

\item{value.call}{A logical value, TRUE if value is function call, FALSE if value is a variable name.}

\item{wait}{See DSI::datashield.aggreate options. Default: FALSE.}

\item{async}{See DSI::datashield.aggreate options. Default: TRUE.}
}
\description{
Cross assign

Cross assign # rather on client side
}
