% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_func.R
\name{crossAssignFunc}
\alias{crossAssignFunc}
\title{Cross assign a preprocessing function}
\usage{
crossAssignFunc(conns, func, symbol)
}
\arguments{
\item{conns}{A list of DSConnection-class.}

\item{func}{Encoded definition of a function for preparation of raw data matrices. 
Two arguments are required: conns (list of DSConnection-classes), 
symbol (name of the R symbol) (see datashield.assign).}

\item{symbol}{Encoded vector of names of R symbols to assign in the Datashield R session on each server in \code{logins}.
The assigned R variables will be used as the input raw data to compute covariance matrix.
Other assigned R variables in \code{func} are ignored.}
}
\description{
Call preprocessing function on remote servers.
}
