% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_func.R
\name{crossProd}
\alias{crossProd}
\title{Matrix cross product}
\usage{
crossProd(x, pair = FALSE, chunk = 500)
}
\arguments{
\item{x}{A list of numeric matrices.}

\item{pair}{A boolean value indicating pairwise cross products are computed. Default, FALSE.}

\item{chunk}{Size of chunks into what the resulting matrix is partitioned. Default: 500.}
}
\value{
A list of cross product matrices
}
\description{
Calculates the cross product of given matrices
}
