% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_func.R
\name{crossProd}
\alias{crossProd}
\title{Matrix cross product}
\usage{
crossProd(x, y = NULL, chunk = 500)
}
\arguments{
\item{x}{A numeric matrix}

\item{y}{A list of numeric matrices. Default, NULL, y = x.}

\item{chunk}{Size of chunks into what the resulting matrix is partitioned. Default: 500.}
}
\value{
\code{t(x) \%*\% y}
}
\description{
Calculates the cross product t(x) \%*\% y
}
