% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_func.R
\name{crossProd}
\alias{crossProd}
\title{Matrix cross product}
\usage{
crossProd(x, y = NULL, pair = FALSE, chunk = 500L)
}
\arguments{
\item{x}{A list of numeric matrices.}

\item{y}{A list of numeric matrices. Default, NULL, \code{y} = \code{x}.}

\item{pair}{A logical value indicating pairwise cross products are computed.
Default, FALSE. Ignored if \code{y} is provided.}

\item{chunk}{Size of chunks into what the resulting matrix is partitioned.
Default, 500L.}
}
\value{
A list of cross product matrices.
}
\description{
Calculates the cross product of given matrices
}
