% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiblock_aggregate.R
\name{crossProd}
\alias{crossProd}
\title{Matrix cross product

Calculates the cross product t(x) \%*\% x}
\usage{
crossProd(x, y = NULL)
}
\arguments{
\item{x}{A numeric matrix}
}
\value{
t(x) \%*\% x
}
\description{
Matrix cross product

Calculates the cross product t(x) \%*\% x
}
