% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{.cleanup}
\alias{.cleanup}
\title{Remove objects from the current workspace}
\usage{
.cleanup(what, start = ".GlobalEnv")
}
\arguments{
\item{what}{a list of  environments and their respective objects -
the output of a previous call to ls.all.}

\item{start}{a character the environment name where to start. Default,
.GlobalEnv.}
}
\description{
This function removes objects from the current workspace.
See \code{dsSwissKnife:::.cleanup}.
}
\keyword{internal}
