% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_func.R
\name{.federateCov}
\alias{.federateCov}
\title{Federated covariance matrix}
\usage{
.federateCov(
  loginFD,
  logins,
  funcPreProc,
  querytables,
  querysubset = NULL,
  pair = FALSE,
  chunk = 500,
  mc.cores = 1,
  connRes = FALSE
)
}
\arguments{
\item{loginFD}{Login information of the FD server (one of the servers containing cohort data)}

\item{logins}{Login information of other servers containing cohort data}

\item{funcPreProc}{Definition of a function for preparation of raw data matrices. 
Two arguments are required: conns (list of DSConnection-classes), 
symbol (name of the R symbol) (see datashield.assign).}

\item{querytables}{Name (or a vector of two names) of the R symbol(s) to assign in the Datashield R session on each server in \code{logins}.
The assigned R variable(s) will be used as the input raw data to compute covariance matrix.
Other assigned R variables in \code{func} are ignored.}

\item{querysubset}{A list of index vectors indicating the subsets of individuals to consider. 
Default, NULL, all individuals are considered.}

\item{pair}{A boolean value indicating pairwise cross products are computed. Default, FALSE.}

\item{chunk}{Size of chunks into what the resulting matrix is partitioned. Default: 500.}

\item{mc.cores}{Number of cores for parallel computing. Default: 1}
}
\value{
Covariance matrix of the virtual cohort
}
\description{
Compute the covariance matrix for the virtual cohort
}
\keyword{internal}
