% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{.ls.all}
\alias{.ls.all}
\title{List all objects in all environments}
\usage{
.ls.all(start = ".GlobalEnv")
}
\arguments{
\item{start}{A character string of the environment (default .GlobalEnv).}
}
\value{
A list of environment names and the respective objects defined in each environment.
}
\description{
\code{ls.all} returns all objects in all environments.
See \code{dsSwissKnife:::.ls.all}.
}
\keyword{internal}
