% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_func.R
\name{.partitionMatrix}
\alias{.partitionMatrix}
\title{Matrix partition}
\usage{
.partitionMatrix(x, seprow, sepcol = seprow)
}
\arguments{
\item{x}{A symmetric matrix}

\item{sep}{A numeric vectors indicating sizes of square blocks}
}
\value{
List of blocks
}
\description{
Partition a symmetric matrix into square blocks
}
\keyword{internal}
