% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_func.R
\name{.rebuildMatrixDsc}
\alias{.rebuildMatrixDsc}
\title{Matrix reconstruction}
\usage{
.rebuildMatrixDsc(dscblocks, mc.cores = 1)
}
\arguments{
\item{dscblocks}{List of lists of bigmemory objects pointed to matrix
blocks.}

\item{mc.cores}{Number of cores for parallel computing. Default, 1.}
}
\value{
The complete symmetric matrix
}
\description{
Rebuild a matrix from its partition bigmemory objects.
}
\keyword{internal}
