% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_func.R
\name{.sumMatrices}
\alias{.sumMatrices}
\title{Sum matrices}
\usage{
.sumMatrices(dsc = NULL, mc.cores = 1)
}
\arguments{
\item{dsc}{A list of big memory descriptions}

\item{mc.cores}{Number of cores for parallel computing. Default: 1}
}
\value{
Sum of matrices stored in dsc
}
\description{
Compute the sum of matrices stored in bigmemory objects
}
\keyword{internal}
