% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serv_scudo.R, R/server_func.R
\name{dsRank}
\alias{dsRank}
\title{Ranking

Ranking of features in each sample}
\usage{
dsRank(x)

dsRank(x)
}
\arguments{
\item{x}{A matrix or data frame, samples in rows and features in columns}
}
\value{
Ranking of features

Ranking of features
}
\description{
Ranking of features in each sample
}
