% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_func.R
\name{rebuildMatrixVar}
\alias{rebuildMatrixVar}
\title{Symmetric matrix reconstruction}
\usage{
rebuildMatrixVar(symbol, len1, len2, mc.cores = 1)
}
\arguments{
\item{symbol}{Generic variable name}

\item{len1}{First-order length of lists of matrix blocks. Variables were generated as symbol__1__1__1, symbol__1__1__2, etc.}

\item{len2}{Second-order length of lists of matrix blocks.}

\item{mc.cores}{Number of cores for parallel computing. Default: 1.}
}
\value{
The complete symmetric matrix
}
\description{
Rebuild a symmetric matrix from its partition through variables in the environment
}
