% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_func.R
\name{singularProd}
\alias{singularProd}
\title{Singular product}
\usage{
singularProd(x)
}
\arguments{
\item{x}{A numeric matrix}
}
\value{
t(x) \%*\% (x \%*\% t(x))[,1]
}
\description{
Product of t(x) and first column of x \%*\% t(x)
}
