% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiblock_aggregate.R
\name{singularProd}
\alias{singularProd}
\title{Product of x' and first column of xx'

Product of x' and first column of xx'}
\usage{
singularProd(x)
}
\arguments{
\item{x}{A numeric matrix}
}
\value{
t(x) %*% (x %*% t(x))[,1]
}
\description{
Product of x' and first column of xx'

Product of x' and first column of xx'
}
