% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_func.R
\name{tcrossProd}
\alias{tcrossProd}
\title{Matrix cross product

Calculates the cross product x \%*\% t(x)}
\usage{
tcrossProd(x, y = NULL)
}
\arguments{
\item{x}{A numeric matrix}
}
\value{
x \%*\% t(x)
}
\description{
Matrix cross product

Calculates the cross product x \%*\% t(x)
}
