% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_func.R
\name{tcrossProd}
\alias{tcrossProd}
\title{Matrix cross product}
\usage{
tcrossProd(x, y = NULL, chunk = 500L)
}
\arguments{
\item{x}{A list of numeric matrices.}

\item{y}{A list of numeric matrices. Default, NULL, y = x.}

\item{chunk}{Size of chunks into what the resulting matrix is partitioned.
Default, 500L.}
}
\value{
\code{x \%*\% t(y)}
}
\description{
Calculates the cross product x \%*\% t(y).
}
