% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiblock_aggregate.R
\name{tripleProd}
\alias{tripleProd}
\title{Matrix triple product

Calculate the triple product}
\usage{
tripleProd(x, y)
}
\arguments{
\item{x}{A numeric matrix}

\item{y}{A list of symmatric numeric matrices of dimension (ncol(x), ncol(x))}
}
\value{
List of x %*% y %*% t(x)
}
\description{
Matrix triple product

Calculate the triple product
}
