% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_func.R
\name{tripleProdChunk}
\alias{tripleProdChunk}
\title{Matrix triple product}
\usage{
tripleProdChunk(x, pids, chunk = 500, mc.cores = 1)
}
\arguments{
\item{x}{A numeric matrix}

\item{pids}{A vector of mate servers' names}

\item{chunk}{Size of chunks into what the resulting matrix is partitioned. Default: 500.}
}
\value{
\code{x \%*\% t(y)}
}
\description{
Calculate the triple product x \%*\% y \%*\% t(x)
}
